Clean up unnecessary calibration plotting
